move url logic into lib.rs file
